- added some code that would allow us to track most calls to the c std library filesystem functions. Running in debug mode the windows taskmanager was showing that egoboo was reading about 10kb a second from some kind of IO. The code replaces the standard fread() function with EGO_fread() and routes this to either the standard fread() function or the custom function _ego_fread() depending on whether the switch DEBUG_STDIO has been defined in egoboo_config.h. This did not isolate the problem so it is not happening through any normal means that I can detect. It could possibly be through the network... This update affected pretty much every file due to the replacement of function names. - added some code to track the number of times a given pip is accessed. The file "pip_usage.txt" will be updated every time that the pips are unloaded (usually when a module is exited).

+++ game/file_formats/mpd_file.h
@@ -102,7 +102,7 @@ struct s_tile_info
     Uint16  img;                               ///< Get texture from this
     Uint8   fx;                                 ///< Special effects flags
     Uint8   twist;
-    Uint32  vrtstart;                           ///< Which vertex to start at
+    size_t  vrtstart;                           ///< Which vertex to start at
 
     bool_t  inrenderlist;
 };
